+++ src/Networking/picoPacket.java
@@ -1,4 +1,4 @@
 package Networking;
 
-public record picoPacket(String type, int err, int pv, int corr) {
+public record picoPacket(String type, double err, double pv, double corr) {
 }
